+++ packages/eslint-plugin-gestalt/src/__fixtures__/no-spread-props/invalid-not-fixable-input.js
@@ -1,9 +1,17 @@
 import { Box, Box as RenamedBox } from 'gestalt';
 
 export default function TestElement(props) {
+  const selectedWrapperProps = {
+    color: 'red',
+    onBlur: props.onBlur,
+    onClick: () => {},
+    onMouseEnter: onMouseEnterHandler,
+    onMouseLeave,
+    rounding: 2,
+  };
 
   return (
-    <Box {...props} width={20}>
+    <Box {...(true ? selectedWrapperProps : {})} width={20}>
       <RenamedBox {...props} width={20} />
     </Box>
   );
