+++ packages/eslint-plugin-gestalt/src/__fixtures__/no-spread-props/valid.js
@@ -1,5 +1,4 @@
 import { Box, Box as RenamedBox } from 'gestalt';
-
 export default function TestElement() {
   return (
     <Box>
