+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/invalid/gestalt-import-with-box-input.js
@@ -1,8 +1,10 @@
+import { useRef } from 'react';
 import { Box } from 'gestalt';
 export default function TestElement() {
+  const ref = useRef(null);
   return (
     <Box>
-      <div ref={undefined}>
+      <div ref={ref}>
         <Box />
       </div>
     </Box>
