+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/invalid/gestalt-import-without-box-input.js
@@ -1,8 +1,9 @@
+import { useRef } from 'react';
 import { Text, Flex } from 'gestalt';
-
 export default function TestElement() {
+  const ref = useRef(null);
   return (
-    <div ref={undefined}>
+    <div ref={ref}>
       <Text>Test</Text>
     </div>
   );
