+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/invalid/gestalt-import-without-box-output.js
@@ -1,8 +1,9 @@
+import { useRef } from 'react';
 import { Box, Flex, Text } from 'gestalt';
-
 export default function TestElement() {
+  const ref = useRef(null);
   return (
-    <Box ref={undefined}>
+    <Box ref={ref}>
       <Text>Test</Text>
     </Box>
   );
