+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/invalid/no-gestalt-import-input.js
@@ -1,6 +1,8 @@
+import { useRef } from 'react';
 export default function TestElement() {
+  const ref = useRef(null);
   return (
-    <div ref={undefined}>
+    <div ref={ref}>
       <div />
     </div>
   );
