+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/invalid/no-gestalt-import-output.js
@@ -1,7 +1,9 @@
 import { Box } from 'gestalt';
+import { useRef } from 'react';
 export default function TestElement() {
+  const ref = useRef(null);
   return (
-    <Box ref={undefined}>
+    <Box ref={ref}>
       <div />
     </Box>
   );
