+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/invalid/no-gestalt-import-selfclosed-input.js
@@ -1,3 +1,5 @@
+import { useRef } from 'react';
 export default function TestElement() {
-  return <div ref={undefined} />;
+  const ref = useRef(null);
+  return <div ref={ref} />;
 }
