+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-box-lonely-ref/valid.js
@@ -1,5 +1,4 @@
-// import { Box } from 'gestalt';
-
+import { Box } from 'gestalt';
 export default function TestElement() {
   return (
     <Box>
