+++ packages/gestalt/src/Button.js
@@ -1,6 +1,5 @@
 // @flow strict
-import type { Node } from 'react';
-import { forwardRef, useImperativeHandle, useRef } from 'react';
+import { forwardRef, useImperativeHandle, useRef, type Node } from 'react';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import Flex from './Flex.js';
